feat: auth 연동을 하였습니다.
issue: #47
