+++ lib/utils/auth.ts
@@ -1,0 +1,4 @@
+import { app } from "./firebaseApp";
+import { getAuth } from "firebase/auth";
+
+export const auth = getAuth(app);
