+++ lib/utils/db.ts
@@ -1,4 +1,5 @@
-import { initializeApp } from "firebase/app";
+import { app } from "./firebaseApp";
+
 import {
   getFirestore,
   collection,
@@ -9,19 +10,6 @@ import {
   query,
   where,
 } from "firebase/firestore";
-
-const firebaseConfig = {
-  apiKey: process.env.API_KEY,
-  authDomain: process.env.AUTH_DOMAIN,
-  projectId: process.env.PROJECT_ID,
-  storageBucket: process.env.STORAGE_BUCKET,
-  messagingSenderId: process.env.MESSAGEING_SENDER_ID,
-  appId: process.env.APP_ID,
-  measurementId: process.env.MEASUREMENT_ID,
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
 
 // Initialize Cloud Firestore and get a reference to the service
 export const db = getFirestore(app);
